requirements.txt
blessed
py-itime
pynvml
psutil
tensorboardX
tensorboard